+++ todo-service/src/main/java/com/m1kellz/todoservice/repository/TodoDetailsRepository.java
@@ -3,5 +3,5 @@ package com.m1kellz.todoservice.repository;
 import com.m1kellz.todoservice.entity.TodoDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TodoDetailsRepository extends JpaRepository<TodoDetails,Long> {
+public interface TodoDetailsRepository extends JpaRepository<TodoDetails,Integer> {
 }
