+++ todo-service/src/main/java/com/m1kellz/todoservice/repository/TodoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TodoRepository extends JpaRepository <Todo,Long> {
-    List<Todo> findAllByUserId(Long userId);
+public interface TodoRepository extends JpaRepository <Todo,Integer> {
+    List<Todo> findAllByUserId(int userId);
 }
